code refactor

+++ src/main/java/com/metric/eticaret/order/service/OrderServiceImpl.java
@@ -1,10 +1,10 @@
 package com.metric.eticaret.order.service;
 
 
-import com.metric.eticaret.authentication.config.JwtTokenUtil;
 import com.metric.eticaret.exception.domain.NotFoundException;
 import com.metric.eticaret.order.model.Order;
 import com.metric.eticaret.order.repository.OrderRepository;
+import com.metric.eticaret.product.repository.ProductRepository;
 import com.metric.eticaret.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.Random;
 
 @Service
@@ -20,24 +21,21 @@ import java.util.Random;
 public class OrderServiceImpl implements OrderService {
 
     private final OrderRepository orderRepository;
-    private final JwtTokenUtil jwtTokenUtil;
+    private final ProductRepository productRepository;
     private final UserRepository userRepository;
 
     @Transactional
     @Override
-    public Order save(Order order) throws NotFoundException {
-
+    public Order save(Order order, Long id) throws NotFoundException {
         if (order != null && order.getId() != null) {
             orderRepository.findById(order.getId()).orElseThrow(() ->
                     new NotFoundException("order not found"));
 
         }
         order.setOrderDate(new Date().getTime());
-        //todo product logic
-       // order.setUser(userRepository.findByUsername(jwtTokenUtil.getUsernameFromToken()));
+        order.setUser(userRepository.findById(id).orElseThrow(()-> new NotFoundException("User not found")));
         Random random = new Random();
-        int orderNumber = random.nextInt(100000);
-        order.setOrderNumber(orderNumber);
+        order.setOrderNumber(random.nextInt(100000));
         BigDecimal totalPrice = order.getPrice().add(order.getPrice().multiply(order.getTaxNumber()));
         totalPrice = totalPrice.add(order.getCargoPrice().subtract((order.getTotalDiscount())));
         order.setTotalPrice(totalPrice);
@@ -51,13 +49,17 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public Order getUser(Long id) throws NotFoundException {
+    public Order getOrder(Long id) throws NotFoundException {
         return orderRepository.findById(id).orElseThrow(() -> new NotFoundException("Order not found"));
     }
 
     @Override
-    public void deleteUser(Long id) {
-        orderRepository.deleteById(id);
+    public void deleteOrderById(Long id) {
+        Optional<Order> orderOptional = orderRepository.findById(id);
+        if (orderOptional.isPresent()){
+            orderRepository.deleteById(id);
+        }
+        //todo id kontrolu
     }
 
 }
